+++ migrators/mitre/mitre_typedb_mapping.py
